+++ src/index.ts
@@ -25,6 +25,7 @@ export function apply(ctx: Context, config: Config = {}): void {
   const telemetry = config.telemetry ?? {
     enabled: process.env.RELAY_PLUGIN_MANAGER_TELEMETRY !== '0',
     endpoint: process.env.RELAY_PLUGIN_MANAGER_TELEMETRY_ENDPOINT,
+    test: process.env.RELAY_PLUGIN_MANAGER_TELEMETRY_TEST === '1',
   }
   ctx.pluginSearch.register(npmSearchProvider())
   ctx.pluginSearch.register(githubSearchProvider())